Move the DisplayWork-creators serialisation tests to the display lib
