+++ sbt_common/display/src/test/scala/uk/ac/wellcome/display/models/DisplayWorkSerialisationTest.scala
@@ -298,4 +298,105 @@ class DisplayWorkSerialisationTest
 
     assertJsonStringsAreEqual(actualJson, expectedJson)
   }
+
+  describe("creators") {
+    it("serialises creators with a mixture of agents/organisations/persons") {
+      val work = IdentifiedWork(
+        canonicalId = "v9w6cz66",
+        sourceIdentifier = sourceIdentifier,
+        version = 1,
+        title = Some("Vultures vying for victory"),
+        creators = List(
+          Unidentifiable(Agent("Vivian Violet")),
+          Unidentifiable(Organisation("Verily Volumes")),
+          Unidentifiable(
+            Person(
+              label = "Havelock Vetinari",
+              prefix = Some("Lord Patrician"),
+              numeration = Some("I")))
+        )
+      )
+      val displayWork = DisplayWork(work)
+
+      val actualJson = objectMapper.writeValueAsString(displayWork)
+      val expectedJson = s"""
+        |{
+        |  "type": "Work",
+        |  "id": "${work.canonicalId}",
+        |  "title": "${work.title.get}",
+        |  "creators": [
+        |    ${identifiedOrUnidentifiable(work.creators(0), abstractAgent)},
+        |    ${identifiedOrUnidentifiable(work.creators(1), abstractAgent)},
+        |    ${identifiedOrUnidentifiable(work.creators(2), abstractAgent)}
+        |  ],
+        |  "subjects": [ ],
+        |  "genres": [ ],
+        |  "publishers": [],
+        |  "placesOfPublication": [ ]
+        |}""".stripMargin
+
+      assertJsonStringsAreEqual(actualJson, expectedJson)
+    }
+
+    it("serialises identified creators") {
+      val work = IdentifiedWork(
+        canonicalId = "v9w6cz66",
+        sourceIdentifier = sourceIdentifier,
+        version = 1,
+        title = Some("Vultures vying for victory"),
+        creators = List(
+          Identified(
+            Person(
+              label = "Havelock Vetinari",
+              prefix = Some("Lord Patrician"),
+              numeration = Some("I")),
+            canonicalId = "hgfedcba",
+            identifiers = List(
+              SourceIdentifier(
+                IdentifierSchemes.libraryOfCongressNames,
+                ontologyType = "Organisation",
+                value = "hv"))
+          ),
+          Identified(
+            Organisation(label = "Unseen University"),
+            canonicalId = "abcdefgh",
+            identifiers = List(
+              SourceIdentifier(
+                IdentifierSchemes.libraryOfCongressNames,
+                ontologyType = "Organisation",
+                value = "uu"))
+          ),
+          Identified(
+            Agent(label = "The Librarian"),
+            canonicalId = "blahbluh",
+            identifiers = List(
+              SourceIdentifier(
+                IdentifierSchemes.libraryOfCongressNames,
+                ontologyType = "Organisation",
+                value = "uu"))
+          )
+        )
+      )
+      val displayWork = DisplayWork(work)
+
+      val actualJson = objectMapper.writeValueAsString(displayWork)
+      val expectedJson = s"""
+                             |{
+                             |  "type": "Work",
+                             |  "id": "${work.canonicalId}",
+                             |  "title": "${work.title.get}",
+                             |  "creators": [
+                             |    ${identifiedOrUnidentifiable(work.creators(0), abstractAgent)},
+                             |    ${identifiedOrUnidentifiable(work.creators(1), abstractAgent)},
+                             |    ${identifiedOrUnidentifiable(work.creators(2), abstractAgent)}
+                             |  ],
+                             |  "subjects": [ ],
+                             |  "genres": [ ],
+                             |  "publishers": [],
+                             |  "placesOfPublication": [ ]
+                             |}""".stripMargin
+
+      assertJsonStringsAreEqual(actualJson, expectedJson)
+    }
+  }
 }
